requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Content & Media
Pillow==10.0.1
django-markdownx==4.0.2
markdown==3.5.1
bleach==6.1.0

# Environment & Configuration
python-decouple==3.8
whitenoise==6.6.0

# Production server
gunicorn==21.2.0

# Development tools
django-debug-toolbar==4.2.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-slugify==8.0.1